requirements.txt
discord.py>=1.3.1
gd.py>=1.0.0b0
import-expression>=1.1.4
